fix: login_required decorator removed from PostDetailView and PostView classes. check_object_permissions added to put and delete methods of PostDetailView in order to check user permission for CRUD operations on post objects

+++ 7/app/views.py
@@ -1,11 +1,9 @@
-from django.shortcuts import render, get_object_or_404
+from django.shortcuts import get_object_or_404
 from rest_framework.views import APIView, Response, status
 
 from .models import Comment, Post
 from .serializers import PostSerializer, CommentSerializer
 
-from django.contrib.auth.decorators import login_required
-from django.utils.decorators import method_decorator
 from .permissions import IsOwnerOrReadOnly
 
 class PostView(APIView):
@@ -15,7 +13,6 @@ class PostView(APIView):
         serializer = PostSerializer(instance=posts, many=True)
         return Response(data=serializer.data, status=status.HTTP_200_OK)
     
-    @method_decorator(login_required)
     def post(self, request):
         serializer = PostSerializer(data=request.data)
         serializer.is_valid(raise_exception=True)
@@ -32,17 +29,17 @@ class PostDetailView(APIView):
         serializer = PostSerializer(instance=post)
         return Response(data=serializer.data, status=status.HTTP_200_OK)
     
-    @method_decorator(login_required)
     def put(self, request, pk):
         post =  get_object_or_404(Post, pk=pk)
-        serializer = PostSerializer(instance=post, data=request.data)
+        self.check_object_permissions(request, post)
+        serializer = PostSerializer(instance=post, data=request.data, partial=True)
         serializer.is_valid(raise_exception=True)
         serializer.save()
         return Response(data=serializer.data, status=status.HTTP_200_OK)
     
-    @method_decorator(login_required)
     def delete(self, request, pk):
         post =  get_object_or_404(Post, pk=pk)
+        self.check_object_permissions(request, post)
         post.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
